Add next.js API endpoints and modify Redux store
In this commit, a significant amount of work has been done to incorporate Next.js API endpoints which are more conducive for server-side rendering, ensuring efficient network fetches. This in particular included new endpoints for verifying the OTP and performing logout operations.

The Redux store has also been expanded to include additional slices, accommodating several new actions and reducers. In the process, 'investorSlice.tsx' was updated to include a reset function, allowing for seamless transition on logout. The 'authSlice' was also similarly updated.

A change in the global CSS was made to correct a padding issue in the auth container for better UX. Additionally, the 'Icon.tsx' was updated to include an additional comma for code consistency.

All in all, these modifications were done with the primary intent to enhance the efficiency and expand the functionality of the application.