Improve data fetching with revalidation
This commit modifies the data fetching function in the dashboard page. The change involves adding revalidation options to the fetch request. The 'revalidate' value, which has been set to 3600 seconds, allows the fetched data to be reused for up to an hour before it is considered stale and a new fetch request is made. This change will improve the performance of the application by reducing the number of unnecessary network requests.