Replace redirect method in otpField component
Changed the method of redirection from `permanentRedirect` to `router.push` in otpField component to ensure the correct post-authentication redirect behavior. This allows us to maintain user's navigation history.

Refactor email/phone validation logic in user auth form

Fixed a logical error in the conditional statement for disabling the button in the user authentication form. Previously, the button was disabled when email input is empty and validation check, but it should be either of them. This fix enables the correct validation.