Fix disabled button logic in auth form
Incorrect condition for the "disabled" prop of the button in the user authorization form was fixed. The button was intended to be disabled if the email was empty OR if certain conditions were false. However, the previous condition was disabling the button if the email was empty AND if certain conditions were false. The updated condition makes sure the user doesn't proceed without entering a valid and non-empty email.