"Add services alias and rename data file in tsconfig.json
A new alias '@/services/*' was added to our tsconfig.json to simplify the import of our services in the store. The '@/data' redirection was also updated to point to index.tsx file from faker.tsx since faker.tsx was explicitly renamed to index.tsx in this commit. This change will make our imports more consistent and clear across the codebase."