Updated startups limit on dashboard
The number of startups in the investor view was increased from 4 to 6 for better data representation. Unnecessary comments were removed and "Model" was imported to facilitate additional functionality. Styling changes were made on the button in the startup component to enhance the user interface. In the liveCampaigns component, only the first 4 campaigns are now displayed for a cleaner arrangement.