Add offline payment & premium membership feature
This code adds a feature for offline payment and indicates premium membership status for users. An 'OfflinePayment' component is created and added to the 'InvestTransactionModal'. Risk acceptance was adjusted to properly reflect the state of 'premiumMember'. In addition, payment processing capabilities were added. As part of this update, the 'InvestForm' file was moved to a new directory. Third-party library 'react-copy-to-clipboard' has been imported to facilitate the copy-to-clipboard feature in offline payments.