Add transactions page and update related components
This commit adds a new TransactionsPage, and corresponding Transaction components to display transactions. A formatCustomDate utility function has also been introduced in utils.ts. Further, updates are made to the Redux configuration and OTP verification functionality in the auth component. Moreover, modifications are done on the Navbar component and NavLink styling, along with the addition of a 'transactions' dropdown item in the UserMenu.