Update import paths in otherDetailsForm component
Imports for Select, Input, UploadCheck and useUpdateContext in otherDetailsForm.tsx file have been updated. The previous paths were leading to incorrect directories which could cause errors. The update corrects these paths to ensure that the components and hook are imported from the correct directories. This allows the form to function as expected.